requirements.txt
streamlit==1.25.0
pandas==1.5.3
numpy==1.23.5
nltk==3.8.1
scikit-learn==1.2.2
Pillow==9.5.0
